% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianMCMC.R
\name{Gaussian.MCMC}
\alias{Gaussian.MCMC}
\title{Gaussian MCMC}
\usage{
Gaussian.MCMC(
  data,
  nrepeat,
  ic = c(k1, k2, k3),
  tun,
  prior.a,
  prior.b,
  Tmax = 100
)
}
\arguments{
\item{data}{input data set. SIR trajectory data set}

\item{nrepeat}{number of iteration of MCMC}

\item{ic}{initial value of beta, gamma, and rho}

\item{tun}{tunning constant for proposal distribution of beta, gamma, rho}

\item{prior.a}{hyper shape parameter of gamma prior for beta, gamma, rho}

\item{prior.b}{hyper rate parameter of gamma prior for beta, gamma, rho}

\item{Tmax}{cutoff time of epidemic}
}
\value{
returning posterior samples of beta, gamma, rho
}
\description{
This function generates posterior samples of beta, gamma, and rho using MCMC based on Gaussian likelihood in subsection A.2
}
