% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SdsMCMC.R
\name{llikelihood}
\alias{llikelihood}
\title{Likelihood function for Eq. (4.3)}
\usage{
llikelihood(SI_ti, p.m, delta.t = delta, n.num = n, nz.num = nz)
}
\arguments{
\item{SI_ti}{return values from SIR.ODE() function.}

\item{p.m}{values of beta, gamma, rho}

\item{delta.t}{duration of infectious period}

\item{n.num}{number of susecptible}

\item{nz.num}{number of removed among initially susceptible individual returning likelihood}
}
\description{
Likelihood function for Eq. (4.3)
}
