% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_ode.R
\name{ode.final.size.cnt}
\alias{ode.final.size.cnt}
\title{ode.final.size.cnt}
\usage{
ode.final.size.cnt(x0 = 100, y0 = 1, gamma = 1, lambda = 1.9, k = 20)
}
\value{

}
\description{
euler.prob
rho=.038
beta=0.2160
gamma=0.1928
}
