% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_ode.R
\name{ode.sir}
\alias{ode.sir}
\title{ode.sir}
\usage{
ode.sir(x, k1 = beta, k2 = gamma)
}
\arguments{
\item{x}{x}

\item{k1}{beta}

\item{k2}{gamma}
}
\value{

}
\description{
euler.prob
rho=.038
beta=0.2160
gamma=0.1928
}
