% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SdsMCMC.R
\name{result}
\alias{result}
\title{MCMC simulation results}
\usage{
result(data = theta, fitn = T, burn1 = burn, nrepeat1 = nrepeat, thin1 = thin)
}
\arguments{
\item{data}{theta - inputted data set. It contains MCMC simulation results.}

\item{fitn}{if T the function estimate initial number of susceptible, if F the function does not estimate N and calculate from data, so inputted data must cover a total population}

\item{burn1}{burn - burning period}

\item{nrepeat1}{nrepeat - number of posterior sample}

\item{thin1}{thin - tinning period}
}
\value{
returning plots that summrizes MCMC simulation results
}
\description{
This Plot summrizes MCMC simulation results.
}
