% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SdsMCMC.R
\name{SDS.Likelihood.MCMC}
\alias{SDS.Likelihood.MCMC}
\title{SDS Likelihood MCMc}
\usage{
SDS.Likelihood.MCMC(
  data = in.data,
  Tmax,
  nrepeat = 1000,
  tun,
  prior.a,
  prior.b,
  ic = c(k1, k2, k3)
)
}
\arguments{
\item{in.data:}{input data set, must be a form of sellke empidemic}

\item{Tmax:}{cutoff time of epidemic}

\item{nrepeat:}{number of iteration of MCMC}

\item{ic:}{initial value of beta, gamma, and rho}

\item{tun:}{tunning constant for proposal distribution of beta, gamma, rho}

\item{prior.a:}{hyper shape parameter of gamma prior for beta, gamma, rho}

\item{prior.b:}{hyper rate parameter of gamma prior for beta, gamma, rho}
}
\value{
returning posterior samples of beta, gamma, rho
}
\description{
This function generates posterior samples of beta, gamma, and rho using MCMC based on SDS likelihood in subsection 4.2
This function uses RAM method via adapt_S() function from ramcmc R-package
}
