% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SdsMCMC.R
\name{SDS.Likelihood.withN.MCMC}
\alias{SDS.Likelihood.withN.MCMC}
\title{SDS Likelihood with N MCMC}
\usage{
SDS.Likelihood.withN.MCMC(
  data = in.data,
  Tmax,
  nrepeat = 1000,
  tun,
  prior.a,
  prior.b,
  ic = c(k1, k2, k3)
)
}
\arguments{
\item{Tmax}{cutoff time of epidemic}

\item{nrepeat}{number of iteration of MCMC}

\item{tun}{tunning constant for proposal distribution of beta, gamma, rho, n}

\item{prior.a}{hyper shape parameter of gamma prior for beta, gamma, rho, hyper paramter lambda}

\item{prior.b}{hyper rate parameter of gamma prior for beta, gamma, rho, hyper paramter lambda}

\item{ic}{initial value of beta, gamma, and rho}

\item{in.data}{input data set, must be a form of sellke empidemic}
}
\value{
returning posterior samples of beta, gamma, rho
}
\description{
This function generates posterior samples of beta, gamma, and rho using MCMC based on SDS likelihood in subsection 4.2
This function includes additional step for generating initial number of susceptible n.
refering to Raftery (1988) , biometiria 75, 2 pp. 223-8
This function uses RAM method via adapt_S() function from ramcmc R-package
}
