% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SdsMCMC.R
\name{SDS.MCMC}
\alias{SDS.MCMC}
\title{SDS Likelihood MCMc}
\usage{
SDS.MCMC(
  data,
  Tmax,
  fitn = T,
  nrepeat = 1000,
  prior.a,
  prior.b,
  ic = c(k1, k2, k3, k4)
)
}
\arguments{
\item{data}{input data set, must be a form of Sellke epidemic (two columns. first is infection time, second is recovery time)}

\item{Tmax}{cutoff time of epidemic}

\item{fitn}{if T the function estimate initial number of susceptible, if F the function does not estimate N and calculate from data, so inputted data must cover a total population}

\item{nrepeat}{number of iteration of MCMC}

\item{prior.a}{hyper shape parameter of gamma prior for beta, gamma, rho, hyper parameter lambda}

\item{prior.b}{hyper rate parameter of gamma prior for beta, gamma, rho, hyper parameter lambda}

\item{ic}{initial value of beta, gamma, rho and N}
}
\value{
returning posterior samples of beta, gamma, rho (with fitn=F) or beta, gamma, rho, and n (with fitn=T)
}
\description{
This function generates posterior samples of beta, gamma, and rho using MCMC based on SDS likelihood in subsection 4.2
This function includes an additional step for generating the initial number of susceptible n. using NegativeBinomial
This function uses RAM method via adapt_S() function from ramcmc R-package
}
