% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sellke.R
\name{Sellke}
\alias{Sellke}
\title{Sellke}
\usage{
Sellke(n, rho = 0.01, beta = 0.3, gamma = 0.15, Tmax = 80, dt = 0.001)
}
\arguments{
\item{n}{number of initial susceptible}

\item{rho}{parmeters, tranmission rate, recovery rate, initial infective ratio}

\item{beta}{parmeters, tranmission rate, recovery rate, initial infective ratio}

\item{gamma}{parmeters, tranmission rate, recovery rate, initial infective ratio}

\item{Tmax}{cutoff time of epidemic}

\item{dt}{time increment}
}
\value{
returns are individual infection time and removed time.

number of rows are n + n*rho
}
\description{
This function generate synthetic epidemic data using Sellke construction in Algorithm 3.1
}
