% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SellkeToTrajecotory.R
\name{Sellke.to.trajectory}
\alias{Sellke.to.trajectory}
\title{Sellke to trajectory}
\usage{
Sellke.to.trajectory(sellke.data, Tmax)
}
\arguments{
\item{sellke.data}{input data set, must be a form of sellke empidemic}

\item{Tmax}{cutoff time of epidemic}
}
\value{
returns are SIR trajectory data set
}
\description{
This function converts epidemic data wih infection and removed time to SIR trajectory data with S(t), I(t), and R(t) at discrete time t
}
