% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SirMle.R
\name{SIR.MLE}
\alias{SIR.MLE}
\title{SIR MLE}
\usage{
SIR.MLE(indata)
}
\arguments{
\item{indata}{input data set. SIR trajectory data set}
}
\value{
returning MLE of beta, gamma, rho of subsection 4.1
}
\description{
This function calculates MLE using SIR empidemic data
}
