% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianMCMC.R
\name{SIR.ODE.full}
\alias{SIR.ODE.full}
\title{This function solves the ODE in Eq. (2.7)}
\usage{
SIR.ODE.full(Tmax, dt = 0.01, beta, gamma, ic)
}
\arguments{
\item{Tmax}{cutoff time of epidemic}

\item{dt}{time increment of ODE}

\item{beta}{present values of beta at the present step at MCMC}

\item{gamma}{present values of gamma at the present step at MCMC}

\item{ic}{initial values of ODE set to c(1.0, rho, 0.0), rho set the present values of beta and gamma at the present step at MCMC}
}
\description{
This function solves the ODE in Eq. (2.7)
}
