% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SdsMCMC.R
\name{SIR.ODE}
\alias{SIR.ODE}
\title{SDS MCMC}
\usage{
SIR.ODE(indata = ti, Tmax = T.max, dt = 0.01, beta, gamma, ic)
}
\arguments{
\item{Tmax}{cutoff time of epidemic}

\item{dt}{time increment of ODE}

\item{beta}{present values of beta at the present step at MCMC}

\item{gamma}{present values of gamma at the present step at MCMC}

\item{ic}{initial values of ODE set to c(1.0, rho, 0.0), rho set the present values of beta and gamma at the present step at MCMC}

\item{inata}{descrete time of an individual. This values are came from Sellke epidemic data}
}
\description{
These functions are to draw posterior samples using MCMC for SDS likelihood in subsection 4.2 and Algorithm 5.1
This function solves the ODE in Eq. (2.7) and step 3 in Algorithm 5.1
}
